Refactor: 2단계 리뷰내용 반영
요청 검증 위치 변경, 도메인 dto 변환로직 위치 변경(toEntity 도입), 테스트 코드 리펙터링
